+++ bee-tangle/benches/tangle_bench.rs
@@ -1,10 +1,10 @@
 // Copyright 2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use bee_message::{Message, MessageId};
+use bee_message::{semantic::ConflictReason, Message, MessageId};
 use bee_runtime::resource::ResourceHandle;
 use bee_storage_null::Storage as NullStorage;
-use bee_tangle::{config::TangleConfig, metadata::MessageMetadata, ConflictReason, Tangle};
+use bee_tangle::{config::TangleConfig, metadata::MessageMetadata, Tangle};
 use bee_test::rand::{message::rand_message, metadata::rand_message_metadata, number::rand_number};
 
 use criterion::*;